Rename the output module to drm
